Refactor score into its own file Bench: 13790343
